+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+## UNRELEASED
+
+- Changed `DynamoDBCacheBackend` to serialize cached value as json string.
+
+
 ## 0.2.0 (2023-09-25)
 
 - Added `CloudflareCacheBackend`.